+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [1.135.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.5...v1.135.0) (2025-07-31)
+
+
+### Features
+
+* support agentic traffic reports for new UI ([#1079](https://github.com/adobe/spacecat-audit-worker/issues/1079)) ([f3c82ef](https://github.com/adobe/spacecat-audit-worker/commit/f3c82efd12c5336f3467588b3b71c8d29eb786ed))
+
+## [1.134.5](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.4...v1.134.5) (2025-07-31)
+
+
+### Bug Fixes
+
+* Upgrade @adobe/structured-data-validator to version 1.2.4 ([#1083](https://github.com/adobe/spacecat-audit-worker/issues/1083)) ([6b88ff8](https://github.com/adobe/spacecat-audit-worker/commit/6b88ff8bd8afa47452d027351452c0ba714d6afa))
+
+## [1.134.4](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.3...v1.134.4) (2025-07-31)
+
+
+### Bug Fixes
+
+* Suggestion error IDs are not unique in Structured Data Audit ([#1077](https://github.com/adobe/spacecat-audit-worker/issues/1077)) ([932cc48](https://github.com/adobe/spacecat-audit-worker/commit/932cc48a3e89cb03d66554e9d44602b14e96f6b8))
+
+## [1.134.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.2...v1.134.3) (2025-07-30)
+
+
+### Bug Fixes
+
+* **metatags:** skip 403/404 pages which were unable to get scraped ([#1060](https://github.com/adobe/spacecat-audit-worker/issues/1060)) ([58ba6ae](https://github.com/adobe/spacecat-audit-worker/commit/58ba6ae31d5bfd110ee24118f294ec40faca3ca5))
+
+## [1.134.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.1...v1.134.2) (2025-07-29)
+
+
+### Bug Fixes
+
+* Suppress ImageObject Error in Structured Data Audit  ([#1066](https://github.com/adobe/spacecat-audit-worker/issues/1066)) ([7a7579d](https://github.com/adobe/spacecat-audit-worker/commit/7a7579d048a724e1099170a17a473d1ee423a030))
+
 ## [1.134.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.134.0...v1.134.1) (2025-07-26)